make a run method
